Added velocity visualization, fixed dynamics1
Dynamics1 had  bug which was that it was only using the
modified acceleration for the position, not for the velocity
calculation. This had some interesting effects on the simulation, in
effect, we had a very good measurement for the velocity.

The bug was discovered once I added in the visualization for the
velocity component (blue dots).